Removes guava dependency
Guava leads to library conflicts and interoperability problems. By
removing the guava dep, we can more safely be placed in core libraries.

Instead of forcing users to use or convert out of Guava's optional, we
now use nullable. Nullable is trivial to convert to an optional type in
the caller's favorite functional library or language.