Removes dependency on commons-lang3 for Base64
This is part of a few changes needed to remove an explicit dependency on
commons-lang3. In this change, we host a local base64 decoder, copied
from okio (under ASL). This decoder is only used once and is 50 loc.

Tests do not have a Java 7 constraint, so can use the built-in Java 8
Base64 codec as opposed to the one from apache commons.