Removes support for Annotation.duration and associated apis
This removes support for Annotation.duration and associated apis. While
this is a breaking change, I can't find any uses even with global GitHub
search, apart from an experiment and spring-cloud-sleuth, which I'll
address.

If we ever want to remove this, it would be best before 3.0 is final.

See https://github.com/openzipkin/zipkin/pull/717