Introduces TraceSampler which replaces TraceFilter
TraceFilter worked off a span id, which isn't always the trace id. This
is confusing and could lead to sparse traces. This eliminates use of
TraceFilter, replaced with TraceSampler, which works off the trace id
and is implemented by a benchmarked implementation in zipkin-java.

The use of zipkin-java's TraceIdSampler is shaded, so as to neither
introduce a dependency, nor the perils of copy/paste.

Impact on users is that they should stop using TraceFilter immediately
as Brave will ignore them. Most often, they will want to only provide a
threshold, ex `TraceSampler.create(0.3f)` means retain 30%.

We also no longer publish `brave-tracefilters` and instead an equivalent
`brave-tracesamplers`, which has the same implementation, except based
on trace id as opposed to span id.

Finally, those using zookeeper should set the rate as a percentage, not
a ratio. For example, `/brave/samplerate` would be 0.2 to retain 20%.