Package required thrift classes for brave-core
The generated model classes require thrift, so thrift is not optional.

libthrift 0.9.0 has a special problem: it includes java source files.
The shade-relocate would extract them and inflate brave-core-shaded to 400k
Newer versions no longer have the problem, and as the model classes were
generated with 0.9.2, I thought this might be a sensible upgrade.