Clean-up README.md
Part of the README.md content was not up to date anymore.
There was information about using Flume as preferred span shipping implementation which is
not the case anymore with HTTP and Kafka support.

Also the overview on how Zipkin/Dapper works doesn't need to be in the brave README.

Also update link to point to OpenZipkin instead of the original Twitter Zipkin link.