Adds HttpSpanCollector.tracingEnabled to prevent trace amplification
Now that Cassandra is instrumentable, we've the possibility of creating
a loop. The scenario is that we are tracing the http collector endpoint,
and the span propagates there. This is helpful as it can help diagnose
performance issues in Zipkin's collector.

The problem happens when cassandra's tracer reports back to the same
http endpoint. This creates a loop, unless we have a flag to disable
sampling.