Moves to zipkin 1.0
zipkin 1.0 is a stable release with semantic version guarantees. This
means we can depend on it directly and not need to shade it.

By not shading it, we have the advantage of other integrations working
without classpath confusion. For example, someone can make a span
collector that writes to zipkin storage directly.