Adds spike of Kafka integration
This abuses local tracer to send a trace across kafka (where kafka
message key is abused to hold propagated state)

There are a number of tradeoffs to consider, not the least of which the
potential of a dual parent on the consumer side. The current
implementation simulates a client-server span which notably includes
boundary annotations of "ms" "mr".

see #164 #211