Make the AnnotationSubmitter.currentTimeMicroseconds() pluggable.
Even in Java8 with java.time the minimum precision for timestamp generation is milliseconds. Providing
microsecond accuracy in spans requires an implementation to do so to be plugged in.

By making the timestamp generation contextual through the AnnotationSubmitter.Clock interface this can be
achieved.

An example implementation to obtain microsecond accuracy using the DataStax CQL java driver
(cassandra-driver-core-*.jar or com.datastax.cassandra:cassandra-driver-core artifact) that uses its AtomicMonotonicTimestampGenerator class is:

    AtomicMonotonicTimestampGenerator TIMESTAMP_GENERATOR = new AtomicMonotonicTimestampGenerator();
    ...
    Brave brave = new Brave
            .Builder("my_service_name")
            .clock(() -> { return TIMESTAMP_GENERATOR.next(); })
            .build();