#264 Make BraveServletFilter customizable
ClientAdapterInterceptor and ServerAdapterInterceptor introduced to make
client and server, request and response adapters customizable.

Most common use case: add further annotations to span

Interceptors makes use of a wrapper approach. The interceptor does not
get in touch with the internals of the various filter implementations.
Instead, he is enabled to intercept and customize any calls to those
implementations. This way the surface area of Brave is kept small.

Two filters (servlet, spring) have been modified to use this approach.
Others to follow once the design is finalized.