Special-case externally provided trace IDs (when sampled left unset)
There are some who want to control trace identifiers, but leave the
sampling decision to the next hop. In this case, they propagate the
required headers (X-B3-TraceId, X-B3-ParentId) and not X-B3-Sampled.

When this happens, the receiver of the headers (the server), should
sample the trace ID before using it.

Before, we weren't implementing this (eventhough Finagle does). The
change here addresses it and also backfills tests accordingly.