Adds Propagation, which decouples B3 implementation from framework code
Before, we had the same algorithm for encoding B3 copy pasted a couple
times. This centralizes it into a new type `Propagation`, and starts the
decoupling of how in-band propagation works from what is reported
out-of-band. The design uses the same terminology as OpenTracing.

This will make B3 maintenance easier as it is in one place. It also
decouples grpc from an incidental dependency on brave-http. Finally, it
allows us to refactor how span data is reported (something orthogonal to
how ids are propagated).

Future change will build on this to make interceptors more flexible,
particularly allowing custom tags to be added via composition as opposed
to inheritance.