Backfills http client instrumentation tests
This backfills http instrumentation tests generically, so that it is
easier to track what supports what (and also how to actually use the
instrumentation!).

Notably, this makes it clear a number of async http clients lose trace
context when their parent is a local span, and that many http clients
don't close spans on exception!