Support Apache HttpAsyncClient "client received" (#308)
Before, BraveHttpResponseInterceptor would call
ClientResponseInterceptor and it would call setClientReceived. But it
would not submit the end annotation because the current client span was
null. This is because the response interceptor is not called on the same
thread as the request interceptor.

This can be solved by storing the client span as an attribute on the
HttpContext instead.