Moves internal utility methods out of Span, adds context and hardening (#311)
The "generated" span type is unfortunately exposed due to the way state
is managed. That said, it is an internal type as all interactions are
recorded via tracers.

Span now has a reference to its propagated context. This prepares for
when we interact with a new version of Brave, as we need a stable lookup
key for mutations such as adding an annotation.

On that note, this eliminates or deprecates places where mutable state
is leaked.