Accepts span.name as empty; asserts against deprecated invocations
Current practice of client spans allows empty string as the span name.
When this is the case, the server's name will be preferred. Before, we
coerced null name to empty, but only in the constructor.

Span shouldn't be created or affected ad-hoc. This hardens span with
assertions as opposed to exceptions as the former won't break production
code.