Locks down internal methods that create spans (#320)
This locks down methods that create `c.t.z.gen.Span`, by forcing them
through the internal package. By doing so, we know if a span was created
in a deprecated way or not. We also remove temptation to use internal
hooks accidentally (by virtue of not reading docs).

This change works by making a package accessor `InternalSpan`, in the
same way as OkHttp works. In the rare cases we need to create spans for
test purposes, we need to set the reference using `initializeForTests()`

The majority of the code impact is old tests which don't need to create
spans explicitly, but would take effort to port to not do so.