Extracts recorder, which centralizes mutation of spans (#322)
This is the last big step in preparation of Brave 4. This adds a package
protected internal type called Recorder, which mimicks the same api in
Brave 4, except that it uses existing logic.

A lot of the work was back-maintenance on tests which in many cases were
redundant or depended on internals.