Prefers end timestamp vs duration in Recorder (#328)
The only user api that specifies a duration is LocalTracer, which is
intended for spans that start and stop without leaving the process.

Since we've recently changed clocks to use an offset approach, we can
get precise duration using start and stop timestamps. We always have
the start timestamp in a local span (as it is required).

With the above in mind, we can safely change internals to use timestamps
to calcuate durations for all spans, not just remote ones. This really
simplifies logic and also makes bridging to OpenTracing easier too, as
its api doesn't accept duration.

The only cost to this approach is short-term, which is that in order
to support the existing LocalTracer.finishSpan(duration), we need to be
able to read-back the span's start timestamp.