Instructions on how to close span on error with JAX-RS + Jersey 2
@lijunyong noticed jax-rs spans don't close on error. This is
problematic as it can leak spans. It is also problematic as spans aren't
highlighted red in the UI.

Unfortunately, there's no hook in JAX-RS for unhandled exceptions. The
only way I can find is to ensure there's always an exception mapper.

Alternatively, we could use framework-specific classes. However, this
would require adding framework-specific modules, which don't yet exist.
If we did that, we'd likely want to do so as a part of #300