Adds brave.Tracing an object graph that holds things you need
Before, we added some helpers like `Tracer.clock()` because when
instrumenting, you sometimes need more than just the Tracer. Recently,
we needed to decide how to expose the propagation format.. should we add
another convenience method?

Taking experience from Zipkin (which took its experience from Dagger and
Spring), another way is to expose an object graph.. a component that
provides types often used together.

This adds `brave.Tracing` which can be passed around or implemented
differently as needed. Instrumentation would consume this to get common
types they need. For example, `JaxRs2Tracing` would accept an instance
of tracing, so that it has what it needs, a tracer and a propagation
handler.

Instances of this can be wired up in Spring (Dagger, Guice or CDI),
making instrumentation wiring far simpler than chaining builders.

Thanks to @bogdandrutu and @llinder for discussions leading to this.