Refines Http client and server handlers and documents (#397)
During integration of the new apis, I noticed some glitches around any
future in sampling (#395) and double-instrumentation (#396). The main
problem was the location of decisions around trace context injection
and extraction, as well when a span was created. The solution to this
problem is to centralize such logic, which has an advantage of making
instrumentation easier to write as well.