Adds brave.Tagger, a safer type to expose for simple customization (#416)
Before, we passed `brave.Span` to callbacks that only add tags. This has
a number of problems, including that it is harder to test and tempts
users to interfere with the Span (like calling finish).

This changes our callbacks to use a simpler type `brave.Tagger`, which
existed in an early version of Brave 4, and recommends exposing this
instead of `brave.Span` for user callbacks like tag customization.