Adds DeclarativeSampler and JAX-RS ContainerAdapter
This adds functionality needed to make sampling decisions based on
annotations present on a Java method (as opposed to runtime parameters).

Specifically, this adds `DeclarativeSampler` which parses a description
of a method into a sample rate (or null to defer the decision).

It also adds `ContainerAdapater`, which can get the JAX-RS
`ResourceInfo` for a method. This supports custom sampling annotations.

Finally, this adds a feature test showing how 3rd-party sampling might
look like using the `DeclarativeSampler` with AspectJ and JAX-RS.