Adds Propagation.Factory.supportsJoin
This adds the ability to inspect if a propagation implementation can
support tracing where client and server share the same span ID. This is
important for alternative mechanisms such as X-Ray or Trace-Context
which do not propagate parent information.

Later pull requests will attempt to integrate another trace propagation
implementation using this.

Fixes #483