Removes expectation a setter can clear a key
I made a mistake documenting Propagation.Setter as being able to clear
a key. This is not guaranteed to work, for example you can't clear
headers made with URLConnection, only replace or add to them.

Instead, this documents that in reusable carriers, you must clear keys
on your own. If carriers are single-shot, there's no need to clear
anything first.