Adds Tracing.Builder.supportsJoin to disable span ID sharing
Similar to ClientServerSameSpan in zipkin-go, this adds the ability to
opt out of span sharing between the client and server side of an RPC.
This is important when reporting to systems that do not share span IDs,
such as Google Stackdriver and Amazon X-Ray.

Note: this still defaults to true, in favor of existing Zipkin
implementations. See https://github.com/openzipkin/zipkin/issues/1480