Deprecates KafkaTracing.joinSpan for nextSpan
`KafkaTracing.joinSpan` will always mutate a completed span when a trace
is in progress. `KafkaTracing.nextSpan` accomplishes this by creating a
separate span for this activity.

The problem with `KafkaTracing.joinSpan` is that it returns a span that
is finished. Any tags or annotations will be delayed until Brave flushes
them:

```
[producer span], [consumer span{end} {late data}]
```

`KafkaTracing.nextSpan` instead returns a child which can be started and
stopped without tainting its parent:

```
[producer span], [consumer span], [processor span]
```