Adds `TraceContext.extra` and supporting command `Tracer.nextSpan` (#514)
This adds the ability to carry extra data from one side of the process
to the other. It is available as `TraceContext.extra`, intentionally
opaque as it is a `Propagation` implementation detail.

For example, a propagation implementation could push extra fields
alongside the trace context, and optionally inject it into outgoing
requests.
```java
ExtraFieldPropagation.newFactory(Propagation.Factory.B3, "user-id");
```

Such a plugin could expose or integrate with other code to make the
extra data visible, given a trace context.

Using the example above, the following is possible using
`Tracing.currentTraceContext()`:
```java
userId = ExtraFieldPropagation.get("user-id");
```

To support this, and to ensure this "extra" data isn't lost, a new
operation is added `Tracer.nextSpan()` which takes the result of
`TraceContext.Extractor.extract()`. Along the way, existing code has
been ported to use it (incidentally fixing bugs not caught before).

A future change can integrate this "extra" with `AWSPropagation` or
expose `ExtraFieldPropagation` if desirable.