Corrects missing status code in servlet 2.5 environments
Before, we couldn't detect the status code unless we wrapped it first.
This made layered tools on servlet api not work unless they could affect
the servlet response. For example, Spring MVC's interceptor cannot.

Benchmarking was used to show that caching the reflective lookup saves
an order of magnitude of overhead than doing so otherwise.