Allows propagation decoration of local trace contexts (#580)
Before, Propagation implementations could not add "extra" data to trace
implementations unless they were extracted from a remote request. This
meant local traces could not propagate extra data.

To fix this, we add a lifecycle hook that decorates the trace context
directly before a span is returned. The first use of this is
`ExtraFieldsPropagation` which can now set fields after a request
started.