Adds http.method tag by default and removes bad naming advice (#616)
This adds the http.method tag by default as many frameworks override
the span name, or will as soon as http.route is supported. This allows
users to always see basic http info at the cost of a small, fixed
cardinality tag. Users who really don't want to see this tag can already
disable parsing of it by overriding the default parser.

This also removes the bad advice from the README, which hinted at using
a path as a span name. This will result in ever-expanding span name
index when there are path variables. The http route based approach is
superior and merges next.