Added an option to wrap current spring rabbit compontents
without this change you can only explicitly create a RabbitTemplate or a SimpleRabbitListenerContainerFactory. In case of frameworks like Sleuth that want to wrap existing instances that's not possible.

with this change in a hacky (RabbitTemplate) and non-hacky (SRLCF) way we're allowing to instrument existing objects. For RT we're hacking into a field of before post processors and injecting there the traced instrumentation. For SRLCF we're adding the traced advice to an existing factory