Clears propagation headers when processing the next Kafka record
Message headers are supposed to represent the parent. A consumer span
should have up to one child, which is a in-process span working on it.
This is similar to a server message being only one span.

This makes the behavior match Rabbit where we remove trace headers when
processing a message.