Adds package-private HttpHandler to optimize current trace context code
Before, we assumed there was no span in scope when calling finish. This
could lead to redundant calls to scope a span (overhead), and also cause
a leak to get to the reporter layer. This centralizes code and tests to
eliminate these problems.

This is particularly important in the work going on for decoupled
context propagation, such as RxJava2 in #665. When others are scoping
spans we don't want to accidentally duplicate scope syncrhonization
overhead.