Ensures http route attribute is written and read on the same thread
This refactors code that reads http route information to do so on
completion of the request. This avoids any thread visibility concerns
such as when request starts and completes on different threads (async).

This also hardens tests which were false positive due to
misconfiguration. Notably, async mode was not enabled on async servlet
tests and the way things were written, the failures went unnoticed.