Introduces SpanRecord type and clarifies a PendingSpan concept
Before, we had some elements that weren't well organized about not-yet-
reported spans. Formerly, there was `MutableSpan`, `MutableSpanMap`, and
`Recorder`. These were internal and replaced by still internal types of:

* SpanRecord: mutable type which can become a Zipkin span.
* PendingSpan: SpanRecord and a clock appropriate for a trace context.
* PendingSpans: repository allowing lookup and removal by trace context.

These are clarifications and optimizations of code that existed before.
Notably, this removes the need to re-build a zipkin span to customize it
before reporting, or to need a zipkin span at all when reporting to
metrics or other aggregations. Future work on that will happen in later
pull requests. There is no public api change involved as all work is
currently internal.