Insulates RxJava2 code from future drift in internal packages (#809)
The "fuseable" package came out in RxJava 2.0 and is as yet still
internal. This puts insulation around it in case it ever becomes not
internal.

Meanwhile there is still risk that some parts of these internal code
drift, for example renaming a method inside ScalarCallable or
ConditionalSubscriber. If this happens, the code will still break.