Update TracingFilter.java
the code is not safe when executed in several threads simultaneously.for example, if "tracer" is null, now threadA execute "setTracing" method, and tracer is not null. but "extractor" and "injector" has not been initialized, just at that moment threadB come, threadB try to execute "invoke" method,because tracer is not null,threadB will go to 72 line or 74 line.because injector and extractor is null,the code will throw NullPointerException,and the code didn't return "invoker.invoke(invocation)".this will cause dubbo service not invoked.
so i use a volatile flag to solve this problem,and make the code thead-safe.