Uses unwrapped type for http parsers and samplers (#990)
While working on Armeria, I noticed that we accidentally missed types
when moving to the standard HTTP client and server responses. What I
mean is that signatures that formerly got an HttpServletRequest now get
an HttpClientRequest. This could break people's parsers without fair
warning.

This makes sure that unwrapped types are used in existing call sites
such as HTTP parsers and samplers. It does so by regearing the new types
as adapters of what they wrap.