Consolidates samplers under the interface SamplerFunction (#998)
This allows more re-use by backporting all single-arity samplers to a
new `SamplerFunction` interface. For example, this consolidates
constants and allows base-layer tools such as `DeclarativeSampler` to
be used in any call site.

This deprecates `Tracer.withSampler` in favor of lighter methods that
accomplish the same:

* `Tracer.startScopedSpan(name, sampler, param)`
* `Tracer.nextSpan(sampler, param)`

With this foundation, we can next implement an RpcRequest type and use
it with less copy/paste of documentation and constants.