Adds basic infrastructure for RPC request sampling
This adds an abstraction layer for RPC, based on non-deprecated code in
brave-instrumentation-http. It will be a work in progress until at least
Dubbo and gRPC are integrated (ideally also Armeria even though in
another repo.

This is limited to sampling logic to make it easy to integrate and also
as that's the more in-demand functionality. Most notably, we only need
two properties: RPC service and method, to satisfy the first use cases
of RPC sampling.