Reverts permittance of duplicate handlers
It is a worse bug to allow duplicate handlers than it is to allow
duplicate configuration. Duplicate handlers translate directly into
runtime overhead, particularly as there is special casing for one
handler.

Throwing an exception was suggested here #961, but that is not a good
choice as it would almost certainly lead to someone having to do an
emergency release to undo it. This changes to logging on dupe, so that
we don't propagate this problem into future work such as collector
handlers.

See https://github.com/openzipkin/zipkin/pull/2807#discussion_r330288383