Adds `RpcTracing` and `HttpTracing.current()` and fixes class loading
When porting Cassandra to `RpcTracing`, I noticed we have no means to
statically look up the tracing instance, such as we do in `Tracing`.
This backfills the same functionality as `brave.Tracing`.

While implementing classloader tests, I notice we load logging types
even if they will be never used. This takes caution to guard usage,
which allows us to test class unloading properly.

To implement this, we had to move the implicit log hook for orphan
tracking to a new option `Tracing.Builder.trackOrphans()`.