Log all exceptions when re-setting message props
When receiving JMS messages from Amazon SQS, the message sometimes
has headers like JMSXGroupID and JMS_SQS_DeduplicationId added by
SQS with empty string values.

When brave-instrumentation-jms is wrapped around an SQSConnectionFactory,
PropertyFilter.filterProperties() attempts to re-set these properties
on the received SQSMessage object. Doing so fails becaus SQSMessage
throws an IllegalArgumentException if either the property name or
value are empty. (Even though the properties came from SQSMessage
to begin with.) The IllegalArgumentException escapes the
try/catch block only looking for JMSException, and results in the
message silently failing to process.

This change widens the scope of the catch block so that any failure
when attempting to re-set a message property is caught, preventing
an unexpected exception type from blowing up the entire message.