WIP: OSGi manifests
I started this and realize a different approach would be better..
Somehow, make a module property, use the same value for OSGi and Java
modules. Make a top-level profile conditional on that property which
adds both. In the case of OSGi, I think since our entries are small,
we can inline them instead of using a separate bnd file