Adds defense for incorrectly implemented client and server requests (#1075)
It is invalid to return null when unwrapping the portable http client
or server types. However, if someone did return null, it is hard to
determine what went wrong. This tolerates the incorrect code similarly
to how we do in toString. The result is that the span will finish and
the code won't crash.