Makes HTTP request method mandatory
While all other properties can be absent or invalid, this makes the HTTP
method mandatory. This will help when we create a path to the request
from the response, as it allows chaining to the method without a null
check.

Note: while it seems path would also always be available even recently,
it has been found missing due to a [request availablility bug](https://github.com/reactor/reactor-netty/pull/998). Eventhough we've seen method incorrectly set in the past,
we've never seen it missing.

Fixes #1089