Hardens HTTP integration tests to catch more bugs easier
It was very hard to detect subtle bugs in reactor until I redid the
tests to force us to be conscious of whether a span error is expected or
if we are expecting a remote or local span. Further, this makes sure
these two most difficult to debug things end up with better assertion
failures, including the entire span json. This is super handy when
tests flake due to timing. For example, using a debugger could actually
slow things to the point they pass!