Works around possibility to have an empty set of extra fields
We didn't eagerly validate `ExtraFieldPropagation`, that it had any
fields at all. This leads to late bugs where an exception is thrown for
trying to inject nothing.

As they type `ExtraFieldPropagation` has public methods in use, this
doesn't change signatures. Rather, it creates a wrapper in the case
nothing exists.

WIP as some unit tests are due.