Adds Tag, Tags and HttpTags
This adds a long overdue feature to ease support of tagging spans. `Tag`
bakes in all logic needed to add a tag to a `Span`, `ScopedSpan`,
`SpanCustomizer` or `MutableSpan` leaving the user left only to decide
what the key is and how to parse it.

This also takes care of error handling and obviation of parsing into a
no-op span.

Here's an example of a potentially expensive tag:
```java
SUMMARY_TAG = new Tag<Summarizer>("summary") {
  @Override protected String parseValue(Summarizer input, TraceContext context) {
    return input.computeSummary();
  }
}

// This works for any variant of span
SUMMARY_TAG.tag(summarizer, span);
```

We need to solve a common case for `BaggageField` which is to add that
as a span tag. This is the best time to fix this problem.

The closest type we have is `ErrorParser` as that does a similar
dispatch. Externally, the closest is [OpenTracing Tag](https://github.com/opentracing/opentracing-java/blob/master/opentracing-api/src/main/java/io/opentracing/tag/Tag.java).

Our approach is different notably as we seal everything except how to
parse the value. This helps avoid leaking exceptions or accidentally
adding null values.