Adds CorrelationScopeDecorator.addDirtyName
`CorrelationScopeDecorator.addDirtyName()` indicates a correlation property
which could be updated without Brave knowing about it. When set, we will
always revert its value on scope close, even if it incurs performance
overhead to track the value.

The value read at the beginning of a scope is currently held when there's a
chance the field can be updated later:

* `CorrelationScopeDecorator.Builder.addDirtyName()`
  * Always revert because someone else could have changed it (ex via `MDC`)
* `BaggageField.flushOnUpdate()`
  * Revert if only when we changed it (ex via `BaggageField.updateValue()`)

This is because users generally expect data to be "cleaned up" when a scope
completes, even if it was written mid-scope.

https://github.com/spring-cloud/spring-cloud-sleuth/issues/1416