Prepares for dynamic baggage
This refactors internal code around baggage so that it can handle
dynamic baggage in a single propagated field. This also removes unused
tag parsing logic in gRPC as we no longer mutate tags.

`BaggageState` is the entry in `TraceContext.extra()` that holds all
baggage, whether they are fixed or not. `BaggageStateHandler` is given
one slot and currently uses one slot per `BaggageField`. However,
`DynamicBaggageStateHandler` shows we can do multiple fields in one
slot, just we aren't doing that yet as we need to define an encoding.