Ensures currentSpan, toSpan and joinSpan use existing context references
Before, `Tracer.currentSpan`, `toSpan` and `joinSpan` either didn't
decorate external contexts (in no-op) or needlessly recreated context
instances when a span was already present in the pending map.i

This not only adds overhead, but interferes with our ability to track
the difference between a newly allocated span vs incidental changes to
an existing one due to use of `currentSpan`, `toSpan` or `joinSpan`.

This change is necessary to support `SpanHandler` when powered by
identity maps.