Extracts WeakConcurrentMap from PendingSpans and uses for trackOrphans
This formalizes that PendingSpans was derived from weaklockfree, pulling
it up so that we can decouple orphan handling in the future.

This also fixes attribution which was in the source, but not NOTICE
before.

Finally, this does a fair amount of simplification on the orphan path.
Most importantly, we pre-emptively fork the context as opposed to
copying most of its values for the orphan path.