Replaces "carrier" with "request"
We prematurely adopted the confusing term "carrier" when trying to work
with OpenTracing years ago. This removes that term for request
everywhere. This does not break any api, as the only thing besides
parameter renames are test methods.

This is in preparation of a future Brave 6 world where ideally no one
will ever need to be confused about what "carrier" is, only for us to
say "an awkward word for request".